requirements.txt
aiohttp==3.6.2
amqp==2.5.2
appdirs==1.4.3
APScheduler==3.6.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
chardet==3.0.4
click==7.1.1
discord.py==1.3.2
entrypoints==0.3
idna==2.9
isort==4.3.21
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-dotenv==0.12.0
pytz==2019.3
regex==2020.2.20
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.0.0
vine==1.3.0
wcwidth==0.1.9
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
